Break up version.go as it contains version and api verification logic
